refactor(uc-scrapper): change structure and modularization
