+++ app/services/uc/scrapper/requests.py
@@ -5,9 +5,7 @@ from urllib.error import HTTPError
 from urllib.parse import urlencode
 
 from app.assets import asset_path
-from app.models.uc import UCSection, UCCourse
-from app.models.base import ClassModule
-from .constants import INFO_INDEX, DAYS
+from .parsers import parse_search
 
 
 def request_table_url(url):
@@ -68,56 +66,6 @@ def request_parameters():
         parameters.append(resource)
 
     return parameters
-
-
-def parse_search(results):
-    courses = {}
-    for line in results:
-        section_html = []
-        for elem in line.find_all("td"):
-            if elem.find_all("table"):
-                aux = []
-                for e in elem.find_all("tr"):
-                    mods = e.find_all("td")
-                    aux.append([m.get_text().replace("\n", "") for m in mods])
-                section_html.append(aux)
-                break
-            else:
-                section_html.append(elem.get_text().replace("\n", ""))
-
-        section = UCSection.get_attributes()
-
-        for attr in INFO_INDEX:
-            attr_info = INFO_INDEX[attr]
-            aux = section_html[attr_info['index']]
-            if aux != "":
-                section[attr] = attr_info['function'](aux.strip())
-
-        for list_ in section_html[-1]:
-            if ":" not in list_[0] or list_[0] == ":":
-                continue  # pragma: no cover
-
-            days_str, modules_str = list_[0].split(":")
-            if days_str == "" or modules_str == "":
-                continue  # pragma: no cover
-
-            for day in days_str.split("-"):
-                for mod in modules_str.split(","):
-                    module = ClassModule.get_attributes()
-                    module['day'] = DAYS[day]
-                    module['module'] = int(mod)
-                    module['type'] = list_[1]
-                    section["modules"].append(module)
-
-        if section["course_code"] not in courses:
-            course = UCCourse.get_attributes()
-            course['name'] = section['name']
-            course['course_code'] = section['course_code']
-            courses[section["course_code"]] = course
-
-        courses[section["course_code"]]["sections"].append(section)
-
-    return list(courses.values())
 
 
 def request_buscacursos(params):
